Extract ClassInfo test builder to a shared trait
